Saving progress (currently compiles, but doesn't run properly).
Adding _kingDir map, which provides the direction to the king of either
color from any square.  Squares that don't have line of sight to a king
have their _kingDir value set to 0.

The _kingDir map is used in conjunction with the _atkd map to quickly
determine pins.  It will also be useful for finding moves that give check
(in QSearchMoveGen).

Also made check evasion move generation cleaner/simpler/faster (well
hopefully it will be faster).